Refactor search index generation to dedicated type
